use channel to pass parsed docs for indexing
