+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "xunmi"
-version = "0.1.0"
+version = "0.2.0"
 edition = "2021"
 license = "MIT"
 documentation = "https://docs.rs/xunmi"
@@ -15,6 +15,7 @@ categories = ["text-processing", "database-implementations"]
 
 [dependencies]
 cang-jie = "0.13.0"
+crossbeam-channel = "0.5"
 fast2s = "0.2"
 jieba-rs = "0.6"
 rayon = "1"
@@ -22,6 +23,7 @@ serde = { version = "1", features = ["derive"] }
 serde-xml-rs = "0.5"
 serde_json = "1"
 serde_yaml = "0.8"
+substring = "1"
 tantivy = "0.16"
 tracing = "0.1"
 tracing-subscriber = "0.3"
