+++ src/lib.rs
@@ -1,10 +1,13 @@
 mod config;
 mod indexer;
+mod input;
 mod updater;
 
 pub use config::{IndexConfig, TextLanguage};
 pub use indexer::Indexer;
-pub use updater::{IndexUpdater, InputConfig, InputType, ValueType};
+pub(crate) use input::Input;
+pub use input::{InputConfig, InputType, ValueType};
+pub use updater::IndexUpdater;
 
 // re-exports
 pub use tantivy::schema::Schema;
@@ -14,14 +17,14 @@ mod tests {
     use serde_json::json;
 
     use super::*;
-    use std::str::FromStr;
+    use std::{str::FromStr, thread, time::Duration};
 
     #[test]
     fn index_data_and_search() {
         let mut config = IndexConfig::from_str(include_str!("../fixtures/config.yml")).unwrap();
         config.path = None;
         let indexer = Indexer::open_or_create(config).unwrap();
-        let mut updater = indexer.get_updater().unwrap();
+        let mut updater = indexer.get_updater();
 
         let data = include_str!("../fixtures/test.yml");
         let config = InputConfig::new(InputType::Yaml, vec![], vec![]);
@@ -39,10 +42,15 @@ mod tests {
         }))
         .unwrap();
         updater.add(&json_data, &config).unwrap();
+        updater.commit().unwrap();
 
-        indexer.reload().unwrap();
+        // need to wait enough time for commit to be ready
+        while indexer.num_docs() == 0 {
+            thread::sleep(Duration::from_millis(100));
+        }
 
         let result = indexer.search("宋元", &["title", "content"], 5, 0).unwrap();
+
         assert_eq!(result.len(), 2);
         let ids: Vec<_> = result
             .iter()
